Update styles, use DatePicker
Using DatePicker for title related dates.  DatePicker returns a DateTime with zeroed out time portion. To ensure consistency in date display, set the flag in Restless Tools at startup so that column displays are not converted to local. Will need to update other parts of the app to use DatePicker and update existing db values.