Update FrameworkElement.Language
Bind FrameworkElement.Language to the current value of xml language. This enables controls (such as DataGrid) to display bound dates in the current culture. Without this, controls don't change, except some specialized ones like DatePicker. ApplicationViewModel handles OnLanguageChanged() to update the CurrentXmlLanguage property, but so far it's not used because would need to call SignalLanguageChange() for a live change. Not ready for that yet.